Windows Classes 2!
This time it actually pushed I have to stop using GitKraken

Major overhaul to how windows are created:
New Window header that contains a 2 classes, one nested within another

'WindowClass' handles the Instance of the window, and 'window' handles the windows API side of the program.

Window.cpp handles everything to do with the instance of the window.